% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/endpoints.R
\docType{package}
\name{appveyor}
\alias{appveyor}
\alias{appveyor-package}
\title{Appveyor API Client}
\description{
This package provides functionality for interacting with the Appveyor API. Appveyor is a continuous integration service that allows for automated testing of software each time that software is publicly committed to a repository on GitHub. Unlike most CIs, it is specifically aimed at a Windows development environment. Setting up Appveyor is quite simple, requiring only a GitHub account, some public (or private) repository hosted on GitHub, and logging into to Appveyor to link it to that repository. Kiril Mueller provides \href{https://github.com/krlmlr/r-appveyor}{straightforward documentation} for configuring your repository to interact with Appveyor via an \samp{appveyor.yml} file added to your repository.

Once you have your Appveyor account configured online, you can use this package to interact with and perform all operations on your Appveyor builds that you would normally perform via the Appveyor website. This includes monitoring builds, modifying build environment settings and environment variables, and cancelling or restarting builds.

Use of this package requires your Appveyor API token, which can be found \href{https://ci.appveyor.com/api-token}{on the Appveyor website}. This should be stored as an environment variable using either \code{Sys.setenv("APPVEYOR_TOKEN" = "exampletoken")} or by storing the API token in your .Renviron file
}
\examples{
\dontrun{
# authenticate using a stored environment variables
Sys.setenv("APPVEYOR_TOKEN" = "appveyorapitoken")

# check to see if you've authenticated correctly
get_user()
}
}

